Abstract out EditPersonDescriptor class
- Moved from EditCommand to TagUtil
- Updated all calls to point to TagUtil
- Basic test cases added
